fix buffer not defined error by adding polyfills
