+++ babel.config.js
@@ -2,4 +2,5 @@ module.exports = {
   presets: [
     '@vue/cli-plugin-babel/preset'
   ],
+
 }
