+++ package.json
@@ -10,25 +10,34 @@
   },
   "dependencies": {
     "1x": "^0.0.0",
+    "@babel/polyfill": "^7.12.1",
+    "@cosmjs/proto-signing": "^0.28.4",
+    "@cosmjs/stargate": "^0.28.6",
     "@polkadot/api": "^8.4.2",
     "@polkadot/extension-dapp": "^0.43.2",
     "@polkadot/ui-keyring": "^2.2.1",
-    "@polkadot/util": "*",
     "@popperjs/core": "^2.11.5",
     "apollo-cache-inmemory": "^1.6.6",
     "apollo-client": "^2.6.10",
     "apollo-link-http": "^1.5.17",
     "axios": "^0.27.2",
+    "bech32": "^2.0.0",
     "blakejs": "^1.2.1",
-    "core-js": "^3.22.8",
+    "core-js": "^3.8.3",
     "country-code-lookup": "^0.0.20",
+    "crypto-browserify": "^3.12.0",
+    "debounce": "^1.2.1",
+    "ethers": "^5.6.8",
     "graphql": "^15.5.1",
     "graphql-tag": "^2.12.6",
     "lodash": "^4.17.21",
+    "long": "^5.2.0",
+    "marked": "^4.0.16",
     "moment": "^2.29.3",
+    "path-browserify": "^1.0.1",
     "qrcode.vue": "1.7.0",
     "stellar-sdk": "^10.1.1",
-    "tfgrid-api-client": "^1.12.0",
+    "stream-browserify": "^3.0.0",
     "vue": "^2.6.14",
     "vue-apollo": "^3.1.0",
     "vue-class-component": "^7.2.3",
@@ -39,13 +48,13 @@
     "vuex": "^3.6.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.5",
     "@babel/plugin-syntax-top-level-await": "^7.14.5",
     "@fortawesome/fontawesome-free": "^6.1.1",
-    "@polkadot/util-crypto": "^9.4.1",
-    "@rollup/plugin-inject": "^4.0.4",
+    "@types/debounce": "^1.2.1",
     "@types/jest": "^27.0.1",
     "@types/lodash": "^4.14.182",
+    "@types/long": "^4.0.2",
+    "@types/marked": "^4.0.3",
     "@typescript-eslint/eslint-plugin": "^5.4.0",
     "@typescript-eslint/parser": "^5.4.0",
     "@vue/cli-plugin-babel": "~5.0.0",
@@ -59,18 +68,13 @@
     "@vue/test-utils": "^1.1.3",
     "@vue/vue2-jest": "^27.0.0-alpha.2",
     "babel-jest": "^27.0.6",
-    "bip39": "^3.0.4",
-    "buffer": "^6.0.3",
-    "crypto": "^1.0.1",
-    "crypto-browserify": "^3.12.0",
     "eslint": "^7.32.0",
     "eslint-plugin-vue": "^8.0.3",
     "jest": "^27.0.5",
-    "path-browserify": "^1.0.1",
-    "process": "^0.11.10",
-    "sass": "^1.52.3",
+    "node-polyfill-webpack-plugin": "^1.1.4",
+    "sass": "^1.32.7",
     "sass-loader": "^12.0.0",
-    "stream-browserify": "^3.0.0",
+    "tfgrid-api-client": "^1.12.1",
     "ts-jest": "^27.0.4",
     "typescript": "~4.5.5",
     "vue-cli-plugin-vuetify": "~2.4.8",
