+++ src/Dashboard.vue
@@ -190,17 +190,7 @@
           {{ new Date().getFullYear() }} — <strong>ThreeFoldTech</strong>
         </v-card-text>
       </v-card>
-      <v-card
-        v-if="$route.path !=='/' && !$route.path.includes('explorer') && !$route.path.includes('hub')"
-        color="#0D47A1"
-        class=" funds px-3 d-flex align-baseline font-weight-bold"
-      > {{balance }} TFT
-        <v-btn
-          @click="addTFT"
-          class="ml-3"
-          :loading="loadingAddTFT"
-        >+</v-btn>
-      </v-card>
+
     </v-footer>
   </v-app>
 </template>
@@ -245,20 +235,10 @@ export default class Dashboard extends Vue {
   $api: any;
   twin: any;
   balance: any = 0;
-  address = "";
+
   accounts: accountInterface[] = [];
   searchTerm = "";
-  loadingAddTFT = false;
-  @Watch("address") async onAddressChanged(value: string, oldValue: string) {
-    if (oldValue.length) {
-      console.log(`removing account ${oldValue}, putting in account ${value}`);
-    } else {
-      console.log(`putting in account ${value}`);
-    }
-  }
-  @Watch("balance") async onBalanceUpdate(value: number, oldValue: number) {
-    console.log(`balance went from ${oldValue}, to ${value}`);
-  }
+
   async created() {
     if (this.$route.path === "/" && !this.$api) {
       Vue.prototype.$api = await connect(); //declare global variable api
@@ -267,22 +247,9 @@ export default class Dashboard extends Vue {
   }
   async mounted() {
     this.accounts = this.$store.state.portal.accounts;
-
-    if (this.$route.path !== "/") {
-      if (this.$route.params.accountID) {
-        this.address = this.$route.params.accountID;
-        this.balance = this.$route.query.balance;
-      }
-    }
   }
   async updated() {
     this.accounts = this.$store.state.portal.accounts;
-    if (this.$route.path !== "/") {
-      if (this.$route.params.accountID) {
-        this.address = await this.$route.params.accountID;
-        this.balance = (await getBalance(this.$api, this.address)) / 1e7;
-      }
-    }
   }
   async unmounted() {
     console.log(`disconnecting from api`);
@@ -300,58 +267,7 @@ export default class Dashboard extends Vue {
     }
     return this.accounts;
   }
-  public async addTFT() {
-    if (config.network !== "dev") {
-      //redirect to https://gettft.com/auth/login?next_url=/gettft/shop/#/buy
-    } else {
-      this.loadingAddTFT = true;
-      getMoreFunds(
-        this.address,
-        this.$api,
-        (res: {
-          events?: never[] | undefined;
-          status: { type: string; asFinalized: string; isFinalized: string };
-        }) => {
-          console.log(res);
-          if (res instanceof Error) {
-            console.log(res);
-            return;
-          }
 
-          const { events = [], status } = res;
-          console.log(`Current status is ${status.type}`);
-          switch (status.type) {
-            case "Ready":
-              this.$toasted.show(`Transaction submitted`);
-          }
-
-          if (status.isFinalized) {
-            console.log(
-              `Transaction included at blockHash ${status.asFinalized}`
-            );
-
-            // Loop through Vec<EventRecord> to display all events
-            events.forEach(({ phase, event: { data, method, section } }) => {
-              console.log(`\t' ${phase}: ${section}.${method}:: ${data}`);
-              if (section === "balances" && method === "Transfer") {
-                this.$toasted.show("Success!");
-
-                getBalance(this.$api, this.address).then((balance) => {
-                  this.balance = balance / 1e7;
-                });
-                this.loadingAddTFT = false;
-              } else if (section === "system" && method === "ExtrinsicFailed") {
-                this.$toasted.show("Get more TFT failed!");
-                this.loadingAddTFT = false;
-              }
-            });
-          }
-        }
-      ).catch((err: { message: string }) => {
-        console.log(err.message);
-      });
-    }
-  }
   public disconnectWallet() {
     this.$store.dispatch("portal/unsubscribeAccounts");
     if (this.$route.query.twinID) {
