+++ src/shims-vue.d.ts
@@ -1,5 +1,6 @@
 declare module "country-iso-3-to-2";
 declare module "tfgrid-api-client";
+
 declare module "*.vue" {
   import Vue from "vue";
   export default Vue;
