+++ vue.config.js
@@ -1,20 +1,17 @@
 const { defineConfig } = require('@vue/cli-service')
-
-
+const NodePolyfillPlugin = require("node-polyfill-webpack-plugin");
 module.exports = defineConfig({
   transpileDependencies: [
     'vuetify'
   ],
-  pluginOptions: 
-   { Buffer: ['buffer', 'Buffer'] , process: "process/browser"}
-  ,
   configureWebpack: {
     resolve: {
       fallback: {
         crypto: require.resolve("crypto-browserify"),
         path: require.resolve("path-browserify"),
         stream: require.resolve("stream-browserify"),
-        buffer: require.resolve("buffer")
       },
-    },},
+    },
+    plugins: [new NodePolyfillPlugin()],
+  },
 })
